chore: changed type to int64

+++ include/aggregation.h
@@ -5,7 +5,7 @@
 #include <vector>
 
 void aggregate_lineitem(const std::shared_ptr<arrow::Table>& lineitem,
-                        const std::unordered_map<int, OrderInfo>& orders_ht,
-                        std::unordered_map<int, AggResult>& agg);
+                        const std::unordered_map<int64_t, OrderInfo>& orders_ht,
+                        std::unordered_map<int64_t, AggResult>& agg);
 
-std::vector<Result> collect_results(const std::unordered_map<int, AggResult>& agg);
+std::vector<Result> collect_results(const std::unordered_map<int64_t, AggResult>& agg);
